extract file list view
#CARPO-30 Done
