+++ src/main/resources/ui/FileList.fxml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import java.lang.*?><?import java.util.*?><?import javafx.scene.*?><?import javafx.scene.control.*?><?import javafx.scene.layout.*?>
+
+<?import com.jfoenix.controls.JFXListView?>
+
+<JFXListView xmlns="http://javafx.com/javafx"
+    xmlns:fx="http://javafx.com/fxml"
+    fx:id="fileList"
+    fx:controller="com.silverhetch.carpo.javafx.FileListView"/>
