+++ src/main/resources/ui/i18n/default.properties
@@ -2,5 +2,7 @@ General.create=Create
 General.search=Search
 General.rename=Rename
 General.rename.hint=Please enter the new name.
+MainView.tabFileList=Files
+MainView.tabTagList=Tags
 MainView.OpenFileFailed=Open file failed.
 DropZone.changePath=...
